requirements.txt
pandas>=2.2.0
openpyxl>=3.1.2
faker>=19.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pycryptodome>=3.20.0
typer[all]>=0.12.0
pyyaml>=6.0.1
tabulate>=0.9.0
Flask>=3.0.0
python-dotenv>=1.0.1
reportlab>=3.6.0

